+++ CHANGELOG.md
@@ -1,2 +1,3 @@
+- Fix education locations
 - Add SP region and municipality type of household
 - Change municipality type to 2014 ARE Gemeindetypologie